+++ tests/factory/test_factory.py
@@ -7,6 +7,7 @@ import edgy
 from edgy.testing import DatabaseTestClient
 from edgy.testing.exceptions import ExcludeValue
 from edgy.testing.factory import FactoryField, ModelFactory
+from edgy.testing.factory.base import ModelFactoryContextImplementation
 from edgy.testing.factory.metaclasses import DEFAULT_MAPPING
 from tests.settings import DATABASE_URL
 
@@ -123,6 +124,96 @@ def test_exclude_autoincrement_factory():
     assert not hasattr(user, "product_ref")
 
 
+def test_sequences():
+    class UserFactory(ModelFactory):
+        class Meta:
+            model = User
+
+        name = FactoryField(
+            callback=lambda field, context, parameters: f"name-{field.get_callcount()}"
+        )
+        product_ref = FactoryField(exclude=True)
+
+    class ProductFactory(ModelFactory):
+        class Meta:
+            model = Product
+
+        user = UserFactory().to_factory_field()
+
+    user = UserFactory().build()
+    assert user.name == "name-1"
+    user = UserFactory().build()
+    assert user.name == "name-2"
+
+    # use the callcounts of the main factory
+    product = ProductFactory().build()
+    assert product.user.name == "name-1"
+    product = ProductFactory().build()
+    assert product.user.name == "name-2"
+
+
+def test_sequences_even():
+    class UserFactory(ModelFactory):
+        class Meta:
+            model = User
+
+        name = FactoryField(
+            callback=lambda field, context, parameters: f"name-{field.inc_callcount()}"
+        )
+        product_ref = FactoryField(exclude=True)
+
+    class ProductFactory(ModelFactory):
+        class Meta:
+            model = Product
+
+        user = UserFactory().to_factory_field()
+
+    user = UserFactory().build()
+    assert user.name == "name-2"
+    user = UserFactory().build()
+    assert user.name == "name-4"
+
+    # use the callcounts of the main factory
+    product = ProductFactory().build()
+    assert product.user.name == "name-2"
+    product = ProductFactory().build()
+    assert product.user.name == "name-4"
+
+
+def test_sequences_odd():
+    class UserFactory(ModelFactory):
+        class Meta:
+            model = User
+
+        name = FactoryField(
+            callback=lambda field, context, parameters: f"name-{field.inc_callcount()}"
+        )
+        product_ref = FactoryField(exclude=True)
+
+    class ProductFactory(ModelFactory):
+        class Meta:
+            model = Product
+
+        user = UserFactory().to_factory_field()
+
+    UserFactory.meta.fields["name"].inc_callcount(
+        amount=-1, callcounts=UserFactory.meta.callcounts
+    )
+    UserFactory.meta.fields["name"].inc_callcount(
+        amount=-1, callcounts=ProductFactory.meta.callcounts
+    )
+    user = UserFactory().build()
+    assert user.name == "name-1"
+    user = UserFactory().build()
+    assert user.name == "name-3"
+
+    # use the callcounts of the main factory
+    product = ProductFactory().build()
+    assert product.user.name == "name-1"
+    product = ProductFactory().build()
+    assert product.user.name == "name-3"
+
+
 def test_exclude_autoincrement_build():
     class UserFactory(ModelFactory):
         class Meta:
@@ -200,7 +291,7 @@ def test_unset():
 
 
 def test_can_use_field_callback_exclude_value():
-    def callback(field_instance, faker, parameters):
+    def callback(field_instance, context, parameters):
         raise ExcludeValue
 
     class ProductFactory(ModelFactory):
@@ -250,17 +341,47 @@ def test_mapping():
             "ManyToMany",
         }:
             DEFAULT_MAPPING[field_type_name](
-                ProductFactory.meta.fields["user"], ProductFactory.meta.faker, {}
+                ProductFactory.meta.fields["user"],
+                ModelFactoryContextImplementation(
+                    faker=ProductFactory.meta.faker,
+                    exclude_autoincrement=ProductFactory.exclude_autoincrement,
+                    depth=0,
+                    callcounts={},
+                ),
+                {},
             )
         elif field_type_name == "ChoiceField":
             DEFAULT_MAPPING[field_type_name](
-                ProductFactory.meta.fields["type"], ProductFactory.meta.faker, {}
+                ProductFactory.meta.fields["type"],
+                ModelFactoryContextImplementation(
+                    faker=ProductFactory.meta.faker,
+                    exclude_autoincrement=ProductFactory.exclude_autoincrement,
+                    depth=0,
+                    callcounts={},
+                ),
+                {},
             )
         elif field_type_name == "RefForeignKey":
             DEFAULT_MAPPING[field_type_name](
-                UserFactory.meta.fields["product_ref"], ProductFactory.meta.faker, {}
+                UserFactory.meta.fields["product_ref"],
+                ModelFactoryContextImplementation(
+                    faker=UserFactory.meta.faker,
+                    exclude_autoincrement=UserFactory.exclude_autoincrement,
+                    depth=0,
+                    callcounts={},
+                ),
+                {},
             )
         else:
             callback = DEFAULT_MAPPING[field_type_name]
             if callback:
-                callback(UserFactory.meta.fields["name"], UserFactory.meta.faker, {})
+                callback(
+                    UserFactory.meta.fields["name"],
+                    ModelFactoryContextImplementation(
+                        faker=UserFactory.meta.faker,
+                        exclude_autoincrement=UserFactory.exclude_autoincrement,
+                        depth=0,
+                        callcounts={},
+                    ),
+                    {},
+                )